remove test_async_executor.py add executor train_from_dataset demo test=develop
